test(source/cloud-storage,tool/cloud-storage-read-object): consolidate tests into single test file per package
Merge TestPageSize into cloudstorage_test.go and TestParseRange into
cloudstoragereadobject_test.go. Both test files now use the internal
package so they can exercise the unexported pageSize and parseRange
helpers directly, removing the need for separate *_internal_test.go
files.
