+++ internal/sources/cloudstorage/cloudstorage_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cloudstorage_test
+package cloudstorage
 
 import (
 	"context"
@@ -21,7 +21,6 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/googleapis/mcp-toolbox/internal/server"
 	"github.com/googleapis/mcp-toolbox/internal/sources"
-	"github.com/googleapis/mcp-toolbox/internal/sources/cloudstorage"
 	"github.com/googleapis/mcp-toolbox/internal/testutils"
 )
 
@@ -40,9 +39,9 @@ func TestParseFromYamlCloudStorage(t *testing.T) {
 			project: my-project
 			`,
 			want: map[string]sources.SourceConfig{
-				"my-gcs": cloudstorage.Config{
+				"my-gcs": Config{
 					Name:    "my-gcs",
-					Type:    cloudstorage.SourceType,
+					Type:    SourceType,
 					Project: "my-project",
 				},
 			},
@@ -99,5 +98,25 @@ func TestFailParseFromYaml(t *testing.T) {
 				t.Fatalf("unexpected error: got %q, want %q", errStr, tc.err)
 			}
 		})
+	}
+}
+
+func TestPageSize(t *testing.T) {
+	tcs := []struct {
+		in   int
+		want int
+	}{
+		{in: 0, want: 1000},
+		{in: -1, want: 1000},
+		{in: 1, want: 1},
+		{in: 500, want: 500},
+		{in: 1000, want: 1000},
+		{in: 1001, want: 1000},
+		{in: 1_000_000, want: 1000},
+	}
+	for _, tc := range tcs {
+		if got := pageSize(tc.in); got != tc.want {
+			t.Errorf("pageSize(%d) = %d, want %d", tc.in, got, tc.want)
+		}
 	}
 }
