--- internal/sources/cloudstorage/pagesize_internal_test.go
@@ -1,37 +1,0 @@
-// Copyright 2026 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package cloudstorage
-
-import "testing"
-
-func TestPageSize(t *testing.T) {
-	tcs := []struct {
-		in   int
-		want int
-	}{
-		{in: 0, want: 1000},
-		{in: -1, want: 1000},
-		{in: 1, want: 1},
-		{in: 500, want: 500},
-		{in: 1000, want: 1000},
-		{in: 1001, want: 1000},
-		{in: 1_000_000, want: 1000},
-	}
-	for _, tc := range tcs {
-		if got := pageSize(tc.in); got != tc.want {
-			t.Errorf("pageSize(%d) = %d, want %d", tc.in, got, tc.want)
-		}
-	}
-}
